requirements/commons.txt
Django==1.8.2
django-celery==3.1.16
SQLAlchemy==1.0.8
django-material==0.4.1
